2_implementation/backend/ai-analysis-service/requirements.txt
# AI Analysis Service Dependencies (MVP minimal set)

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7

# Google Gemini SDK
google-generativeai==0.8.5

# Database
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.6
redis==5.0.7
rq==1.16.2